Allow custom rust target to be used
The existing config structure is not flexible enough to allow multiple
rust targets for the same architecture. This allows specifying the
target explicitly (also includes a cli parameter that can optionally
used by the users of the library). If not provided, it will fallback to
the target in the config.