Remove rustflags for apple builds
Until now rust flags for apple devices were being set on the
GLOBAL_CONFIG dictionary. This disallow the developer to customize
them as overrides this value when set on the manifest file.

This commit removes this value from the GLOBAL_CONFIG, so the
projects who were relying on it as default will have to set them on
their manifest file.