Fix paths for dylibs for Apple platforms
Embedded paths in dylibs are absolute paths of
where they were built. This makes it so that on CI
it embeds the full CI path of artefacts which is then
traversed and the binary is not found when linking.

To fix that `install_name_tool` must be used.

The effects can be observed by using MacOS tool
called `otool`(part of XCode tools).

```
$ otool lib.dylib
...
Load command 3
    cmd LC_ID_DYLIB
    name /Users/[REDACTED]/target/lib.dylib
...
```

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>