Fix OpenWRT images
protoc was unecessarily installed in /root/
which required permissions, however at the end
it switched the user back to "buildbot" without changing
the installed "protoc" binary permissions.

The fix seems to be easy - just remove the root
altogether and install it at "/builder" which is owned by
"buildbot" user already.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>